requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
dvc~=2.5.4
awscli
flake8
python-dotenv>=0.5.1
pandas~=1.1.5
numpy~=1.19.5
PyYAML>=5.4
setuptools~=51.3.3
scikit-learn~=0.24.0
plotly-express~=0.4.0
hyperopt~=0.2.5
kaggle~=1.5.12
pytest
tableone~=0.7.10
xgboost~=1.3.3
